requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pandas
azure-storage-blob==12.8.1
openpyxl==3.0.7
opencv-python
onnxruntime
pillow
azure-servicebus
sqlalchemy
pymysql
pyjwt
scipy
Pillow